pyproject.toml
[project]
name = "parrotlabs"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "category-encoders>=2.9.0",
    "matplotlib>=3.10.8",
    "pandas>=3.0.1",
    "torch>=2.6.0",
    "lightning>=2.5.0",
    "hydra-core>=1.3.0",
    "omegaconf>=2.3.0",
    "xgboost>=2.1.0",
    "optuna>=4.0.0",
    "scikit-learn>=1.5.0",
    "scipy>=1.15.0",
    "tensorboard>=2.19.0",
    "pydantic>=2.7.0",
    "pyyaml>=6.0.3",
    "torchvision>=0.25.0",
    "pyarrow>=23.0.1",
    "shap>=0.51.0",
]

[tool.setuptools.packages.find]
where = ["."]

[dependency-groups]
dev = [
    "ipykernel>=7.2.0",
    "pytest>=8.0.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]
